# backend/database/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type AuthCode struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	CodeHash   string
	ExpiresAt  pgtype.Timestamptz
	ConsumedAt pgtype.Timestamptz
	Attempts   int32
	CreatedAt  pgtype.Timestamptz
}

type User struct {
	ID        uuid.UUID
	Email     string
	CreatedAt pgtype.Timestamptz
}

type WaterLog struct {
	ID              uuid.UUID
	UserID          uuid.UUID
	AmountMl        int32
	LoggedAt        pgtype.Timestamptz
	IsDeleted       bool
	ServerUpdatedAt pgtype.Timestamptz
}