# shared-schemas/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Local-first sync engine.",
    "title": "Drinkwater Sync API",
    "contact": {},
    "version": "1.0"
  },
  "basePath": "/",
  "paths": {
    "/auth/me": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Returns the account for the supplied bearer token. Used by clients to validate a stored token on startup.",
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Get the current user",
        "responses": {
          "200": {
            "description": "The authenticated user",
            "schema": {
              "$ref": "#/definitions/api.User"
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal server or database error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/auth/request": {
      "post": {
        "description": "Sends a 6-digit login code to the given email, creating the account if it does not yet exist. Always returns a generic acknowledgement so the endpoint cannot be used to probe which emails have accounts.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Request a login code",
        "parameters": [
          {
            "description": "Email to send the login code to",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api.AuthRequestInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generic acknowledgement",
            "schema": {
              "$ref": "#/definitions/api.MessageResponse"
            }
          },
          "400": {
            "description": "Invalid request body or email",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "429": {
            "description": "A code was requested too recently",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal server or database error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "description": "Exchanges a valid email + 6-digit code for a 30-day session token. Wrong guesses are capped per code; a successful verification consumes the code so it cannot be reused.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Verify a login code",
        "parameters": [
          {
            "description": "Email and the 6-digit code",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/api.AuthVerifyInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session token and the authenticated user",
            "schema": {
              "$ref": "#/definitions/api.AuthResponse"
            }
          },
          "400": {
            "description": "Invalid request body or email",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Code is wrong, expired, used up, or locked",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal server or database error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/logs": {
      "get": {
        "description": "Returns the non-deleted water logs whose logged_at falls within the half-open range [from, to). The client passes the start of the selected local day and the start of the following day so day boundaries honour the client's timezone. Used to view historical days that are no longer cached locally.",
        "produces": ["application/json"],
        "tags": ["logs"],
        "summary": "Fetch water logs for a specific day",
        "parameters": [
          {
            "type": "string",
            "description": "ISO-8601 timestamp marking the inclusive start of the range",
            "name": "from",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "ISO-8601 timestamp marking the exclusive end of the range",
            "name": "to",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Logs within the requested range, most recent first",
            "schema": {
              "$ref": "#/definitions/api.LogsResponse"
            }
          },
          "400": {
            "description": "Missing or invalid range parameters",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal server or database error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/sync": {
      "post": {
        "description": "Pushes local changes (upserts) to the server and pulls remote changes (delta) made on other devices.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["sync"],
        "summary": "Sync local water logs with the server",
        "parameters": [
          {
            "type": "string",
            "description": "ISO-8601 Timestamp of the client's last sync",
            "name": "since",
            "in": "query"
          },
          {
            "description": "Array of local water logs to sync",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/api.WaterLog"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful sync response containing delta changes",
            "schema": {
              "$ref": "#/definitions/api.SyncResponse"
            }
          },
          "400": {
            "description": "Invalid request body",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal server or database error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "api.AuthRequestInput": {
      "type": "object",
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "api.AuthResponse": {
      "type": "object",
      "required": ["token", "user"],
      "properties": {
        "token": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/api.User"
        }
      }
    },
    "api.AuthVerifyInput": {
      "type": "object",
      "required": ["code", "email"],
      "properties": {
        "code": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "api.LogsResponse": {
      "type": "object",
      "required": ["logs"],
      "properties": {
        "logs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api.WaterLog"
          }
        }
      }
    },
    "api.MessageResponse": {
      "type": "object",
      "required": ["message"],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "api.SyncResponse": {
      "type": "object",
      "required": ["changes", "server_time"],
      "properties": {
        "changes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api.WaterLog"
          }
        },
        "server_time": {
          "type": "string"
        }
      }
    },
    "api.User": {
      "type": "object",
      "required": ["email", "id"],
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "api.WaterLog": {
      "type": "object",
      "required": ["amount_ml", "id", "is_deleted", "logged_at"],
      "properties": {
        "amount_ml": {
          "type": "integer"
        },
        "id": {
          "type": "string"
        },
        "is_deleted": {
          "type": "boolean"
        },
        "logged_at": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and the session token returned by /auth/verify.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}